requirements.txt
# Основные зависимости
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Для нечеткого сравнения строк
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Для работы с переменными окружения
python-dotenv>=1.0.0

# Для Telegram бота
python-telegram-bot>=20.0

# Дополнительно (опционально)
tqdm>=4.65.0  # для прогресс-баров
colorama>=0.4.6  # для цветного вывода в консоль